chore: allow denom to be returned from keeper
