+++ middleware.go
@@ -1,0 +1,3 @@
+package dollar
+
+// TODO(@justin): Populate this with the IBC Middleware for blocking $USDN transfers via ICS 20.
